:recycle: views_common 분리
